fix: update services
