+++ lib/services/services.dart
@@ -1,0 +1,3 @@
+export 'match_service.dart';
+export 'message_service.dart';
+export 'user_service.dart';
